tests/test_full_receipt_flow.py
"""
Step 24: End-to-End Testing
============================
PROMPT Reference: Phase 9, Step 24

Complete workflow tests covering:
    1. Stub upload → OCR → inventory update
    2. Telegram upload → processed <3 sec
    3. Gemini OCR accuracy (5 receipts, 90%+ target)
    4. Gemini rate-limit → Ollama fallback
    5. MQTT sync <2 sec
    6. Home Assistant dashboard updates
    7. Recommendation generation
    8. Spending calculations
    9. Backup/restore
    10. Offline mode (disconnect → reconnect)
"""

import os
import json
import pytest
import logging

logger = logging.getLogger(__name__)

# ---------------------------------------------------------------------------
# Test Configuration
# ---------------------------------------------------------------------------

BACKEND_URL = os.getenv("BACKEND_URL", "http://localhost:8080")
TEST_RECEIPTS_DIR = os.path.join(os.path.dirname(__file__), "test_receipts")


# ---------------------------------------------------------------------------
# Test 1: Stub Upload → OCR → Inventory
# ---------------------------------------------------------------------------

class TestReceiptUpload:
    """Test the stub upload endpoint processes receipts correctly."""

    def test_upload_valid_receipt(self):
        """Upload a receipt image and verify OCR processes it."""
        # TODO: Implement
        # response = requests.post(
        #     f"{BACKEND_URL}/receipts/upload",
        #     headers={"Authorization": "Bearer TEST_TOKEN"},
        #     files={"image": open("test_receipt.jpg", "rb")},
        # )
        # assert response.status_code in (200, 202)
        # data = response.json()
        # assert "status" in data
        pytest.skip("Not yet implemented")

    def test_upload_invalid_file(self):
        """Upload a non-image file and verify graceful error."""
        # TODO: Implement
        pytest.skip("Not yet implemented")

    def test_upload_without_auth(self):
        """Upload without Bearer token and verify 401."""
        # TODO: Implement
        pytest.skip("Not yet implemented")


# ---------------------------------------------------------------------------
# Test 2: OCR Accuracy
# ---------------------------------------------------------------------------

class TestOCRAccuracy:
    """Test OCR extraction accuracy across multiple receipts."""

    def test_gemini_accuracy(self):
        """Verify Gemini OCR achieves 90%+ accuracy on 5 test receipts."""
        # TODO: Implement with real receipt images
        pytest.skip("Not yet implemented — requires test receipt images")

    def test_ollama_fallback(self):
        """Force Gemini failure and verify Ollama processes successfully."""
        # TODO: Implement
        pytest.skip("Not yet implemented")


# ---------------------------------------------------------------------------
# Test 3: MQTT Sync
# ---------------------------------------------------------------------------

class TestMQTTSync:
    """Test real-time MQTT sync latency."""

    def test_inventory_update_sync(self):
        """Verify inventory update appears via MQTT within 2 seconds."""
        # TODO: Implement
        pytest.skip("Not yet implemented")


# ---------------------------------------------------------------------------
# Test 4: Recommendations
# ---------------------------------------------------------------------------

class TestRecommendations:
    """Test deal and seasonal recommendation generation."""

    def test_deal_detection(self):
        """Verify deal detection with price variations."""
        # TODO: Implement
        pytest.skip("Not yet implemented")

    def test_seasonal_detection(self):
        """Verify seasonal pattern detection."""
        # TODO: Implement
        pytest.skip("Not yet implemented")


# ---------------------------------------------------------------------------
# Test 5: Spending Analytics
# ---------------------------------------------------------------------------

class TestAnalytics:
    """Test spending calculation accuracy."""

    def test_monthly_spending(self):
        """Verify monthly spending matches manual calculation."""
        # TODO: Implement
        pytest.skip("Not yet implemented")

    def test_budget_alert(self):
        """Verify 80% budget alert triggers correctly."""
        # TODO: Implement
        pytest.skip("Not yet implemented")


# ---------------------------------------------------------------------------
# Test 6: Backup & Restore
# ---------------------------------------------------------------------------

class TestBackupRestore:
    """Test backup creation and restore."""

    def test_backup_creates_archive(self):
        """Verify backup script creates a valid archive."""
        # TODO: Implement
        pytest.skip("Not yet implemented")

    def test_restore_recovers_data(self):
        """Verify restore brings back all data."""
        # TODO: Implement
        pytest.skip("Not yet implemented")
